# lib/api-zod/src/generated/types/pairRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PairRequest {
  /** Phone number with country code (e.g. 2547xxxxxxxx) */
  phone: string;
}