# lib/api-zod/src/generated/types/pairResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PairResponseStatus } from "./pairResponseStatus";

export interface PairResponse {
  /** The pairing code (e.g. XXXX-XXXX-XXXX) */
  code: string;
  /** Unique session identifier to poll for completion */
  sessionId: string;
  status: PairResponseStatus;
}