# lib/api-zod/src/generated/types/sessionStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionStatusStatus } from "./sessionStatusStatus";

export interface SessionStatus {
  sessionId: string;
  status: SessionStatusStatus;
  /** The captured session ID string (prefixed with TRUTH_MD_CREDS;) when ready */
  sessionData?: string;
}